# backend/middleware-protocolli/api-docs.json
{"swagger":"2.0","info":{"description":"Microservizio Middleware","version":"1.0","title":"Bari Città Connessa Middleware"},"host":"localhost:8080","basePath":"/","tags":[{"name":"protocollo-controller","description":"Protocollo Controller"}],"paths":{"/middleware/v1/protocollo/numero/{number}/anno/{year}":{"get":{"tags":["protocollo-controller"],"summary":"getProtocollo","operationId":"getProtocolloUsingGET","produces":["application/json"],"parameters":[{"name":"number","in":"path","description":"number","required":true,"type":"string"},{"name":"year","in":"path","description":"year","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GetProtocolloWebResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/middleware/v1/protocollo/richiesta/entrata":{"post":{"tags":["protocollo-controller"],"summary":"richiestaProtocollo","operationId":"richiestaProtocolloUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"protocolloEntrataRequest","description":"protocolloEntrataRequest","required":true,"schema":{"$ref":"#/definitions/ProtocolloEntrataWebRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ProtocolloEntrataWebResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/middleware/v1/protocollo/richiesta/entrata_new":{"post":{"tags":["protocollo-controller"],"summary":"richiestaEntrata_new","operationId":"richiestaEntrata_newUsingPOST","consumes":["multipart/form-data"],"produces":["application/json"],"parameters":[{"name":"protocolloRequest.allegati[0].documento.titolo","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.mittente.personaGiuridica.ragioneSociale","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.mittente.pecEmail","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.contenuto","in":"formData","required":false,"type":"string","format":"byte"},{"name":"protocolloRequest.allegati[0].documento.improntaMIME.value","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.allegati[0].documento.dettaglio","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.destinatari[0].personaFisica.nome","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.allegati[0].documento.improntaMIME.algoritmo","in":"formData","required":false,"type":"string"},{"name":"allegati","in":"formData","required":false,"type":"array","items":{"type":"file"},"collectionFormat":"multi"},{"name":"protocolloRequest.mittente.nazione","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.sunto","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.allegati[0].documento.tipoRiferimento","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.destinatari[0].personaFisica.cognome","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.improntaMIME.value","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.allegati[0].documento.improntaMIME.codifica","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.allegati[0].documento.classifica","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.classifica","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.destinatari[0].personaFisica.codiceFiscale","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.mittente.email","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.improntaMIME.algoritmo","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.mittente.personaFisica.nome","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.allegati[0].documento.nomeFile","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.mittente.indirizzo","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.mittente.personaFisica.codiceFiscale","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.oggetto","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.destinatari[0].amministrazione","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.tipoRiferimento","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.titolo","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.destinatari[0].areaOrganizzativaOmogenea","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.allegati[0].documento.sunto","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.improntaMIME.codifica","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.mittente.personaGiuridica.partitaIVA","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.idUtente","in":"formData","required":false,"type":"integer","format":"int32"},{"name":"protocolloRequest.allegati[0].documento.collocazioneTelematica","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.dettaglio","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.mittente.personaFisica.cognome","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.mittente.comune","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.areaOrganizzativaOmogenea","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.allegati[0].documento.contenuto","in":"formData","required":false,"type":"string","format":"byte"},{"name":"protocolloRequest.amministrazione","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.collocazioneTelematica","in":"formData","required":false,"type":"string"},{"name":"protocolloRequest.documento.nomeFile","in":"formData","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/middleware/v1/protocollo/richiesta/uscita":{"post":{"tags":["protocollo-controller"],"summary":"richiestaProtocolloUscita","operationId":"richiestaProtocolloUscitaUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"protocolloUscitaWebRequest","description":"protocolloUscitaWebRequest","required":true,"schema":{"$ref":"#/definitions/ProtocolloUscitaWebRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ProtocolloUscitaWebResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/middleware/v1/protocollo/richiesta/uscita_new":{"post":{"tags":["protocollo-controller"],"summary":"richiestaProtocolloUscita_new","operationId":"richiestaProtocolloUscita_newUsingPOST","consumes":["multipart/form-data"],"produces":["application/json"],"parameters":[{"name":"protocolloUscitaRequest.mittente.uo","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.documento.improntaMIME.codifica","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].modalitaSpedizione","in":"formData","required":false,"type":"integer","format":"int32"},{"name":"protocolloUscitaRequest.allegati[0].documento.collocazioneTelematica","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.email.oggetto","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.allegati[0].documento.contenuto","in":"formData","required":false,"type":"string","format":"byte"},{"name":"protocolloUscitaRequest.documento.improntaMIME.algoritmo","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.documento.titolo","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].nazione","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.documento.dettaglio","in":"formData","required":false,"type":"string"},{"name":"allegati","in":"formData","required":false,"type":"array","items":{"type":"file"},"collectionFormat":"multi"},{"name":"protocolloUscitaRequest.amministrazione","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.documento.classifica","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.allegati[0].documento.improntaMIME.codifica","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.documento.sunto","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.allegati[0].documento.classifica","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].idProvincia","in":"formData","required":false,"type":"integer","format":"int32"},{"name":"protocolloUscitaRequest.mittente.cognome","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.allegati[0].documento.sunto","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.documento.contenuto","in":"formData","required":false,"type":"string","format":"byte"},{"name":"protocolloUscitaRequest.documento.nomeFile","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.email.testo","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.documento.improntaMIME.value","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.allegati[0].documento.improntaMIME.value","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].ragioneSociale","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.allegati[0].documento.improntaMIME.algoritmo","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].nome","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].comune","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.email.tipoAllegati","in":"formData","required":false,"type":"integer","format":"int32"},{"name":"protocolloUscitaRequest.allegati[0].documento.tipoRiferimento","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].cognome","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].indirizzo","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.documento.collocazioneTelematica","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.areaOrganizzativaOmogenea","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].piva","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.allegati[0].documento.dettaglio","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.mittente.nome","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.allegati[0].documento.nomeFile","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].fax","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.documento.tipoRiferimento","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].email","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.allegati[0].documento.titolo","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.idUtente","in":"formData","required":false,"type":"integer","format":"int32"},{"name":"protocolloUscitaRequest.destinatari[0].codiceFiscale","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.destinatari[0].pecEmail","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.email.from","in":"formData","required":false,"type":"string"},{"name":"protocolloUscitaRequest.oggetto","in":"formData","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ProtocolloUscitaWebResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Allegato":{"type":"object","properties":{"documento":{"xml":{"name":"documento","attribute":false,"wrapped":false},"$ref":"#/definitions/Documento"}},"title":"Allegato","xml":{"name":"Allegato","attribute":false,"wrapped":false}},"ContattoDestinatario":{"type":"object","properties":{"codiceFiscale":{"type":"string"},"cognome":{"type":"string"},"comune":{"type":"string"},"email":{"type":"string"},"fax":{"type":"string"},"idProvincia":{"type":"integer","format":"int32"},"indirizzo":{"type":"string"},"modalitaSpedizione":{"type":"integer","format":"int32"},"nazione":{"type":"string"},"nome":{"type":"string"},"pecEmail":{"type":"string"},"piva":{"type":"string"},"ragioneSociale":{"type":"string"}},"title":"ContattoDestinatario","xml":{"name":"ContattoDestinatario","attribute":false,"wrapped":false}},"DatiEmail":{"type":"object","properties":{"from":{"type":"string"},"oggetto":{"type":"string","xml":{"name":"oggetto","attribute":false,"wrapped":false}},"testo":{"type":"string","xml":{"name":"testo","attribute":false,"wrapped":false}},"tipoAllegati":{"type":"integer","format":"int32"}},"title":"DatiEmail","xml":{"name":"DatiEmail","attribute":false,"wrapped":false}},"Destinatario":{"type":"object","properties":{"amministrazione":{"type":"string","xml":{"name":"amministrazione","attribute":false,"wrapped":false}},"areaOrganizzativaOmogenea":{"type":"string","xml":{"name":"areaOrganizzativaOmogenea","attribute":false,"wrapped":false}},"personaFisica":{"xml":{"name":"personaFisica","attribute":false,"wrapped":false},"$ref":"#/definitions/PersonaFisica"}},"title":"Destinatario","xml":{"name":"Destinatario","attribute":false,"wrapped":false}},"Documento":{"type":"object","properties":{"classifica":{"type":"string"},"collocazioneTelematica":{"type":"string"},"contenuto":{"type":"string","format":"byte","xml":{"name":"contenuto","attribute":false,"wrapped":false}},"dettaglio":{"type":"string"},"improntaMIME":{"$ref":"#/definitions/ImprontaMIME"},"nomeFile":{"type":"string","xml":{"name":"nomeFile","attribute":false,"wrapped":false}},"sunto":{"type":"string"},"tipoRiferimento":{"type":"string","xml":{"name":"tipoRiferimento","attribute":true,"wrapped":false}},"titolo":{"type":"string","xml":{"name":"titolo","attribute":false,"wrapped":false}}},"title":"Documento","xml":{"name":"Documento","attribute":false,"wrapped":false}},"Errore":{"type":"object","properties":{"codice":{"type":"string","xml":{"name":"codice","attribute":false,"wrapped":false}},"descrizione":{"type":"string","xml":{"name":"descrizione","attribute":false,"wrapped":false}}},"title":"Errore","xml":{"name":"Errore","attribute":false,"wrapped":false}},"GetProtocolloWebResponse":{"type":"object","properties":{"return":{"$ref":"#/definitions/Return"}},"title":"GetProtocolloWebResponse","xml":{"name":"getProtocolloResponse","namespace":"http://server.ws.protocollo.linksmt.it/","attribute":false,"wrapped":false}},"ImprontaMIME":{"type":"object","properties":{"algoritmo":{"type":"string","xml":{"name":"algoritmo","attribute":true,"wrapped":false}},"codifica":{"type":"string","xml":{"name":"codifica","attribute":true,"wrapped":false}},"value":{"type":"string"}},"title":"ImprontaMIME","xml":{"name":"ImprontaMIME","attribute":false,"wrapped":false}},"Mittente":{"type":"object","properties":{"comune":{"type":"string"},"email":{"type":"string"},"indirizzo":{"type":"string"},"nazione":{"type":"string"},"pecEmail":{"type":"string"},"personaFisica":{"$ref":"#/definitions/PersonaFisica"},"personaGiuridica":{"$ref":"#/definitions/PersonaGiuridica"}},"title":"Mittente","xml":{"name":"Mittente","attribute":false,"wrapped":false}},"MittenteProtUscita":{"type":"object","properties":{"cognome":{"type":"string"},"nome":{"type":"string"},"uo":{"type":"string"}},"title":"MittenteProtUscita","xml":{"name":"MittenteProtUscita","attribute":false,"wrapped":false}},"PersonaFisica":{"type":"object","properties":{"codiceFiscale":{"type":"string"},"cognome":{"type":"string","xml":{"name":"cognome","attribute":false,"wrapped":false}},"nome":{"type":"string","xml":{"name":"nome","attribute":false,"wrapped":false}}},"title":"PersonaFisica","xml":{"name":"PersonaFisica","attribute":false,"wrapped":false}},"PersonaGiuridica":{"type":"object","properties":{"partitaIVA":{"type":"string","xml":{"name":"partitaIVA","attribute":false,"wrapped":false}},"ragioneSociale":{"type":"string","xml":{"name":"ragioneSociale","attribute":false,"wrapped":false}}},"title":"PersonaGiuridica","xml":{"name":"PersonaGiuridica","attribute":false,"wrapped":false}},"ProtocolloEntrataWebRequest":{"type":"object","properties":{"protocolloRequest":{"$ref":"#/definitions/ProtocolloRequest"}},"title":"ProtocolloEntrataWebRequest","xml":{"name":"richiestaProtocollo","namespace":"http://server.ws.protocollo.linksmt.it/","attribute":false,"wrapped":false}},"ProtocolloEntrataWebResponse":{"type":"object","properties":{"return":{"$ref":"#/definitions/Return"}},"title":"ProtocolloEntrataWebResponse","xml":{"name":"richiestaProtocolloResponse","namespace":"http://server.ws.protocollo.linksmt.it/","attribute":false,"wrapped":false}},"ProtocolloRequest":{"type":"object","properties":{"allegati":{"type":"array","xml":{"name":"allegati","attribute":false,"wrapped":false},"items":{"$ref":"#/definitions/Allegato"}},"amministrazione":{"type":"string","xml":{"name":"amministrazione","attribute":false,"wrapped":false}},"areaOrganizzativaOmogenea":{"type":"string","xml":{"name":"areaOrganizzativaOmogenea","attribute":false,"wrapped":false}},"destinatari":{"type":"array","xml":{"name":"destinatari","attribute":false,"wrapped":false},"items":{"$ref":"#/definitions/Destinatario"}},"documento":{"xml":{"name":"documento","attribute":false,"wrapped":false},"$ref":"#/definitions/Documento"},"idUtente":{"type":"integer","format":"int32"},"mittente":{"xml":{"name":"mittente","attribute":false,"wrapped":false},"$ref":"#/definitions/Mittente"},"oggetto":{"type":"string","xml":{"name":"oggetto","attribute":false,"wrapped":false}}},"title":"ProtocolloRequest","xml":{"attribute":false,"wrapped":false}},"ProtocolloUscitaRequest":{"type":"object","properties":{"allegati":{"type":"array","xml":{"name":"allegati","attribute":false,"wrapped":false},"items":{"$ref":"#/definitions/Allegato"}},"amministrazione":{"type":"string","xml":{"name":"amministrazione","attribute":false,"wrapped":false}},"areaOrganizzativaOmogenea":{"type":"string","xml":{"name":"areaOrganizzativaOmogenea","attribute":false,"wrapped":false}},"destinatari":{"type":"array","xml":{"name":"destinatari","attribute":false,"wrapped":false},"items":{"$ref":"#/definitions/ContattoDestinatario"}},"documento":{"xml":{"name":"documento","attribute":false,"wrapped":false},"$ref":"#/definitions/Documento"},"email":{"$ref":"#/definitions/DatiEmail"},"idUtente":{"type":"integer","format":"int32"},"invio":{"type":"boolean"},"mittente":{"$ref":"#/definitions/MittenteProtUscita"},"oggetto":{"type":"string","xml":{"name":"oggetto","attribute":false,"wrapped":false}},"riservato":{"type":"boolean"}},"title":"ProtocolloUscitaRequest","xml":{"attribute":false,"wrapped":false}},"ProtocolloUscitaWebRequest":{"type":"object","properties":{"protocolloUscitaRequest":{"$ref":"#/definitions/ProtocolloUscitaRequest"}},"title":"ProtocolloUscitaWebRequest","xml":{"name":"richiestaProtocolloUscita","namespace":"http://server.ws.protocollo.linksmt.it/","attribute":false,"wrapped":false}},"ProtocolloUscitaWebResponse":{"type":"object","properties":{"return":{"$ref":"#/definitions/Return"}},"title":"ProtocolloUscitaWebResponse","xml":{"name":"richiestaProtocolloUscitaResponse","namespace":"http://server.ws.protocollo.linksmt.it/","attribute":false,"wrapped":false}},"QName":{"type":"object","properties":{"localPart":{"type":"string"},"namespaceURI":{"type":"string"},"prefix":{"type":"string"}},"title":"QName"},"Return":{"type":"object","properties":{"allegati":{"type":"array","xml":{"name":"allegati","attribute":false,"wrapped":false},"items":{"$ref":"#/definitions/Allegato"}},"amministrazione":{"type":"string"},"areaOrganizzativaOmogenea":{"type":"string"},"dataProtocollo":{"$ref":"#/definitions/XMLGregorianCalendar"},"destinatari":{"type":"array","xml":{"name":"destinatari","attribute":false,"wrapped":false},"items":{"$ref":"#/definitions/Destinatario"}},"documento":{"$ref":"#/definitions/Documento"},"errore":{"$ref":"#/definitions/Errore"},"idUtente":{"type":"integer","format":"int32"},"mittente":{"$ref":"#/definitions/Mittente"},"numeroProtocollo":{"type":"integer","format":"int64"},"oggetto":{"type":"string"}},"title":"Return","xml":{"attribute":false,"wrapped":false}},"XMLGregorianCalendar":{"type":"object","properties":{"day":{"type":"integer","format":"int32"},"eon":{"type":"integer","format":"int64"},"eonAndYear":{"type":"integer","format":"int64"},"fractionalSecond":{"type":"number"},"hour":{"type":"integer","format":"int32"},"millisecond":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"month":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"timezone":{"type":"integer","format":"int32"},"valid":{"type":"boolean"},"xmlschemaType":{"$ref":"#/definitions/QName"},"year":{"type":"integer","format":"int32"}},"title":"XMLGregorianCalendar"}}}