# backend/middleware-protocolli/src/main/java/it/almaviva/baricittaconnessaprotocollomiddleware/data/richiestaprotocollo/web/GetProtocolloWebResponse.java
package it.almaviva.baricittaconnessaprotocollomiddleware.data.richiestaprotocollo.web;


import it.almaviva.baricittaconnessaprotocollomiddleware.data.richiestaprotocollo.soap.GetProtocolloResponse;

public class GetProtocolloWebResponse extends GetProtocolloResponse {
}